shift component fixs
